+++ src/App.vue
@@ -2,14 +2,15 @@
 - src/App.vue
 -
 - The frame of the application - same for all pages
+-
+- Reference:
+-   - "Error tracking with Vue.js" (blog, Aug 2018)
+-     -> https://rollbar.com/blog/error-tracking-vue-js/
 -->
 <template>
-  <!--
-  <fatal v-if="fatalMessage"/>
-  -->
   <header>
     <app-logo />
-    <div id="mode" v-bind:class="{ devLocal: mode === 'dev_local', devOnline: mode === 'development' }">
+    <div id="mode" v-bind:class="{ devLocal: mode === 'dev_local' /*, devOnline: mode === 'development'*/ }">
     </div>
     <!-- Note: 'user' can be 'null' at first, then either an object (signed in) or 'false' (signed out).
     -->
@@ -20,31 +21,41 @@
   </main>
   <footer>
     <app-footer />
+    <button id="errorBtn" v-if="mode !== 'production'" v-on:click="makeError">Make error!</button>
   </footer>
 </template>
 
-<style scoped>
+<style scoped lang="scss">
+  header {
+    position: relative;   /* allow 'fatal' to be totally on its own (children can use 'position:relative') */
+  }
   #mode {
-    padding: 10px;
-    position: fixed;
+    position: relative;
     top: 40px;
     left: 0;
+    width: 9.1em;   /* Q: how to make it automatically scale, based on contents (from CSS)? #help #css */
+    padding: 10px;
     display: none;
+    border-radius: 10px;
   }
   #mode.devLocal {
     display: block;
     background-color: dodgerblue;
-    content: 'EMULATION MODE';
+    &:after {
+      content: 'EMULATION MODE';
+    }
   }
-  #mode.devOnline {
-    display: none;
-    background-color: indianred;
-    content: 'ONLINE MODE';
+
+  #errorBtn {
+    position: fixed;
+    bottom: 0;
+    right: 0;
+    margin: 0.3em;
   }
 </style>
 
 <script>
-  import { onMounted } from 'vue'
+  import { onMounted, getCurrentInstance, ref } from 'vue'
 
   import AppLogo from './components/AppLogo.vue'
   import AppProfile from './components/AppProfile/index.vue'
@@ -55,13 +66,63 @@
   // Start maintaining 'userInfo' collection when the logged in user changes
   import './firebase/userInfo'
 
-  function setup() {
+  import { vueWarning } from "./logging";
+
+  function setup(props) {
+    const { mode } = props;
+
+    console.debug("App mode:", mode);
+
     // Note: We rather take the title from here than in 'public/index.html', keeping it application agnostic.
     onMounted(() => {
       console.log("Houston, App is mounted");
     });
 
-    return { user }
+    const appConfig = getCurrentInstance().appContext.config;
+      //
+      assert (appConfig.errorHandler === undefined);    // we're not overwriting anything
+      assert (appConfig.warnHandler === undefined);
+
+    /***  We benefit from this only if we wish to see Vue specific context.
+    // '.errorHandler' catches errors "during component rendering" (well, and watch callbacks, lifecycle hooks,
+    //                  component event handlers) - but still only Vue specific code.
+    //
+    // err: "error trace"
+    // vm: "component in which error occurred"
+    // info: "setup function" |...
+    //    "Vue specific error information such as lifecycle hooks, events etc."
+    //
+    appConfig.errorHandler = (err, vm, info) => {   // (Error, Proxy, string) => ()
+
+      throw err;    // pass on
+    }
+    **/
+
+    // For warnings, most important to see them in the browser console (disabled in production, by Vue itself)
+    //
+    if (mode !== 'production') {   // "only works during development"
+      appConfig.warnHandler = (msg, vm, trace) => {
+        console.warn("Vue warning:", {msg, vm, trace});
+
+        central(vueWarning, `Vue warning: ${msg}`, {vm, trace});
+      }
+    }
+
+    // Note: We get the Vue warning:
+    //  <<
+    //    Property "user" was accessed during render but is not defined on instance.
+    //  <<
+    //
+    //   This is not harmful but a way to mitigate such warning is appreciated. #help
+
+    function makeError() {
+      throw new Error("Error for testing!");
+    }
+
+    return {
+      user,
+      makeError
+    }
   }
 
   export default {
@@ -70,7 +131,7 @@
       AppLogo, AppProfile, AppFooter
     },
     props: {
-      mode: String    // 'dev_local', 'development' or 'production'
+      mode: String    // 'dev_local'|'development'|'production'
     },
     setup
   }