+++ src/assert.js
@@ -5,8 +5,8 @@
 *
 * The alternative would be to a) not use assert, b) use 'rollup-plugin-node-builtins' but brings in a lot more, as well.
 *
-* tbd. Compare this with node.js 'assert' source and maybe learn from there. E.g. can we make the stack trace point to
-*     the caller of 'assert'? #contribute
+* tbd. Compare this with node.js 'assert' source and maybe learn from there.
+* tbd. Can we make the stack trace point to the caller of 'assert'? #contribute
 */
 function assert(cond, msgOpt) {
   if (!cond) {